Move the lexer into own crate
