+++ .gitignore
@@ -1,4 +1,4 @@
-/target
+**/target
 .hypothesis
 *.profraw
