+++ Cargo.toml
@@ -1,7 +1,19 @@
+[workspace]
+members = ["dtl-lexer"]
+resolver = "3"
+
+[workspace.package]
+edition = "2024"
+
+[workspace.dependencies]
+miette = { version = "7.6.0", features = ["fancy"] }
+thiserror = "2.0.16"
+num-bigint = "0.4.6"
+
 [package]
 name = "django-rusty-templates"
 version = "0.1.0"
-edition = "2024"
+edition.workspace = true
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
@@ -13,15 +25,15 @@ cached = "0.56.0"
 either = "1.15.0"
 encoding_rs = "0.8.35"
 html-escape = "0.2.13"
-miette = { version = "7.6.0", features = ["fancy"] }
-num-bigint = "0.4.6"
+miette.workspace = true
+num-bigint.workspace = true
 pyo3 = { version = "0.27.0", features = ["num-bigint"] }
 sugar_path = "1.2.0"
-thiserror = "2.0.16"
-unicode-xid = "0.2.6"
+thiserror.workspace = true
 regex = "1.11.2"
 unicode-normalization = "0.1.24"
 num-traits = "0.2.19"
+dtl-lexer = { path = "dtl-lexer" }
 
 [dev-dependencies]
