+++ dtl-lexer/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "dtl-lexer"
+version = "0.1.0"
+edition.workspace = true
+
+[dependencies]
+miette.workspace = true
+thiserror.workspace = true
+num-bigint.workspace = true
+unicode-xid =  "0.2.6"
