+++ dtl-lexer/src/custom_tag.rs
@@ -2,11 +2,11 @@ use miette::{Diagnostic, SourceSpan};
 use thiserror::Error;
 use unicode_xid::UnicodeXID;
 
-use crate::lex::common::{
+use crate::common::{
     LexerError, NextChar, lex_numeric, lex_text, lex_translated, lex_variable, text_content_at,
     translated_text_content_at,
 };
-use crate::lex::tag::TagParts;
+use crate::tag::TagParts;
 use crate::types::TemplateString;
 
 #[derive(Debug, PartialEq)]
