+++ dtl-lexer/src/lib.rs
@@ -6,6 +6,7 @@ pub mod forloop;
 pub mod ifcondition;
 pub mod load;
 pub mod tag;
+pub mod types;
 pub mod variable;
 
 pub const START_TAG_LEN: usize = 2;
